moved swipeable to wrap the graph only
