+++ src/views/CompoundFinance/CompoundFinance.js
@@ -7,7 +7,6 @@ import { CompoundContext } from '../../App'
 import Grid from '@material-ui/core/Grid'
 import { makeStyles } from '@material-ui/core/styles'
 import useMediaQuery from '@material-ui/core/useMediaQuery';
-import { Swipeable } from 'react-swipeable';
 
 import Container from '../../components/Container'
 import Hero from '../../components/Hero'
@@ -51,17 +50,6 @@ const CompoundFinance = () => {
   const smallScreen = !useMediaQuery('(min-width:1280px)');
   const mobileView = !useMediaQuery('(min-width:600px)'); // This has to match xs, sm, etc...
 
-  // http://stack.formidable.com/react-swipeable/
-  // Add some styling + animation later
-  const swipeRight = () => currentProposal > 1 ? setCurrentProposal(currentProposal-1) : null;
-  const swipeLeft = () => currentProposal < proposals.length ? setCurrentProposal(currentProposal + 1) : null;
-  const config = {
-    delta: 10,
-    preventDefaultTouchmoveEvent: true,
-    trackTouch: true,
-    trackMouse: true,
-  }
-
   return (
     <>
       <Hero proposal={proposal} />
@@ -83,13 +71,7 @@ const CompoundFinance = () => {
               />
             </Grid>
             <Grid item xs={12} lg={6} style={classes.mainGraph} style={{order: smallScreen ? '-1' : null}}>
-              <Swipeable
-                onSwipedRight={swipeRight}
-                onSwipedLeft={swipeLeft}
-                {...config}
-              >
-                <MainPanel proposal={proposal} smallScreen={smallScreen} />
-              </Swipeable>
+              <MainPanel proposal={proposal} smallScreen={smallScreen} />
             </Grid>
             <Grid item xs={12} sm={6} lg={3}>
               <VotesPanel
